[Fix/#122] 온보딩(현 회원가입) 시간표 DTO로 변환 안 되는 문제 해결
